NET-276 Article number needs to be long too!
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/net/trunk@1080844 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/net/nntp/NNTPClient.java
@@ -227,7 +227,7 @@ public class NNTPClient extends NNTP
         if (parts.length > 6) {
             int i = 0;
             try {
-                article.setArticleNumber(Integer.parseInt(parts[i++]));
+                article.setArticleNumber(Long.parseLong(parts[i++]));
                 article.setSubject(parts[i++]);
                 article.setFrom(parts[i++]);
                 article.setDate(parts[i++]);
